to get uptodate with the NBs (1.3)

+++ branches/r1.2/src/CacheWolf/Preferences.java
@@ -47,9 +47,9 @@ import ewe.util.Comparer;
 import ewe.util.Enumeration;
 import ewe.util.Hashtable;
 import ewe.util.Iterator;
-import ewe.util.Map.MapEntry;
 import ewe.util.StringTokenizer;
 import ewe.util.Utils;
+import ewe.util.Map.MapEntry;
 import ewesoft.xml.MinML;
 import ewesoft.xml.sax.AttributeList;
 
@@ -331,6 +331,7 @@ public class Preferences extends MinML {
 	public int maxLogsToKeep = DEFAULT_MAX_LOGS_TO_SPIDER;
 	/** keep own logs even when excessing <code>maxLogsToKeep</code> */
 	public boolean alwaysKeepOwnLogs = true;
+	public boolean overwriteLogs = false;
 
 	/** Determines whether to fill the white areas on the map */
 	public boolean fillWhiteArea = false;
@@ -605,6 +606,7 @@ public class Preferences extends MinML {
 		} else if (name.equals("spider")) {
 			doNotGetFound = !Boolean.valueOf(atts.getValue("getFinds")).booleanValue();
 			checkLog = Boolean.valueOf(atts.getValue("checkLog")).booleanValue();
+			overwriteLogs  = Boolean.valueOf(atts.getValue("overwriteLogs")).booleanValue();
 			tmp = atts.getValue("checkTBs");
 			if (tmp != null)
 				checkTBs = Boolean.valueOf(atts.getValue("checkTBs")).booleanValue();
@@ -857,6 +859,7 @@ public class Preferences extends MinML {
 			outp.print("    <spider" //
 					+ " spiderUpdates=\"" + SafeXML.strxmlencode(spiderUpdates) + "\"" //
 					+ " checkLog=\"" + SafeXML.strxmlencode(checkLog) + "\"" //
+					+ " overwriteLogs=\"" + SafeXML.strxmlencode(overwriteLogs) + "\"" //
 					+ " checkTBs=\"" + SafeXML.strxmlencode(checkTBs) + "\"" //
 					+ " checkDTS=\"" + SafeXML.strxmlencode(checkDTS) + "\"" //
 					+ " spiderRoute=\"" + SafeXML.strxmlencode(spiderRoute) + "\"" //
@@ -1207,21 +1210,31 @@ public class Preferences extends MinML {
 			new MessageBox(MyLocale.getMsg(321, "Error"), MyLocale.getMsg(ErrorMsgActive, "[Profile active...]"), FormBase.MBOK).execute();
 		} else {
 			boolean err = true;
+			FileBugfix profilePath = new FileBugfix(absoluteBaseDir + f.newSelectedProfile);
 			if (operation == 3) {
 				String newName = new InputBox("Bitte neuen Verzeichnisnamen eingeben : ").input("", 50);
 				if (!newName.equals(null)) {
-					err = !renameDirectory(new FileBugfix(absoluteBaseDir + f.newSelectedProfile), new FileBugfix(absoluteBaseDir + newName));
+					err = !renameDirectory(profilePath, new FileBugfix(absoluteBaseDir + newName));
 				}
 			} else if (operation == 2) {
 				Profile p = new Profile();
 				p.dataDir = absoluteBaseDir + f.newSelectedProfile + "/";
 				p.readIndex();
 				String mapsPath = absoluteBaseDir + "maps" + p.getRelativeCustomMapsPath();
-				int answer = new MessageBox("", mapsPath + " " + MyLocale.getMsg(143, "l?schen ?"), FormBase.MBYESNO).execute();
+				//Really check if the user wants to delete the profile
+				String questionText = MyLocale.getMsg(276, "Do You really want to delete profile '")+
+						f.newSelectedProfile + MyLocale.getMsg(277, "' ?");
+				int answer = new MessageBox("", questionText, FormBase.MBYESNO).execute();
+				if (answer != 1){
+					//No? Exit!
+					return;
+				}
+
+				answer = new MessageBox("", mapsPath + " " + MyLocale.getMsg(143, "l?schen ?"), FormBase.MBYESNO).execute();
 				if (answer == 1) {
 					deleteDirectory(new FileBugfix(mapsPath));
 				}
-				err = !deleteDirectory(new FileBugfix(absoluteBaseDir + f.newSelectedProfile));
+				err = !deleteDirectory(profilePath);
 				// ? wait until deleted ?
 			}
 			if (err) {
